refactor : getProDuct의 쿼리 스트링 생성 함수 유틸로 분리
